+++ src/index.ts
@@ -1,2 +1,6 @@
+/* istanbul ignore file */
+
 export * from './AbstractModel';
 export * from './Property';
+export * from './PropertyOptions';
+export * from './ExposedPropertyMetadata';