+++ test/fixtures/SomeClass.ts
@@ -1,5 +1,7 @@
-import { AbstractModel } from "../../src/AbstractModel";
-import { Property } from "../../src/Property";
+/* istanbul ignore file */
+
+import { AbstractModel } from '../../src/AbstractModel';
+import { Property } from '../../src/Property';
 
 export class SomeClass extends AbstractModel<SomeClass> {
     @Property()