+++ test/fixtures/circular-reference/user.ts
@@ -1,0 +1,12 @@
+/* istanbul ignore file */
+
+import { Property } from '../../../src';
+import { UserInfo } from './user-info';
+
+export class User {
+    @Property()
+    public id: number;
+
+    @Property()
+    public userInfo: UserInfo;
+}